fix: PCA-Verbindungsflag, EK-HCL-Rundung, Function-Property-Info-Leak
- LEDModule::checkI2cConnection: setzt bei Verbindungsverlust pcaI2cConnection
  = false. Vorher blieb das Flag true -> bis zu 10 s Schreibzugriffe an einen
  toten PCA9685 und das Frontpanel meldete faelschlich "Dimmer: OK".
- DimChannel_EK::setHcl: round(brightness * 2.55) statt round((uint)(...)) -
  der Cast vor round hatte einen 1-LSB-Floor-Bias.
- LEDModule::processFunctionProperty case 2 (unimplementiert): setzt jetzt eine
  definierte, leere Antwort (resultData[0]=0, resultLength=1) statt mit
  'return true' bis zu 255 Byte uninitialisierten Stack ueber den KNX-Bus
  zurueckzusenden (Info-Leak aus dem Security-Review).

+++ src/DimChannel_EK.cpp
@@ -184,7 +184,7 @@ void DimChannel_EK::setHcl(uint8_t channel, uint16_t kelvin, uint8_t brightness)
     if (ParamEK_hclStart == PT_hclStart_during) { // HCL active when channel is on
         if (ParamEK_hclCheckBrightness == 1 && _isOn) {
             logDebugP("HCL active - Channel: %i Kelvin: %i Brightness: %i", channel, kelvin, brightness);
-            _newValueEK = round((uint)(brightness * 2.55));
+            _newValueEK = round(brightness * 2.55); // 0-100 -> 0-255 (round auf den float, nicht cast-vor-round)
             startTask(DimTaskEK::EK_DIM_B_SET);
         } else {
             _currentHclValue[0] = brightness;
